Update exceptions handlers & changes in service impl-intf

+++ ms-ticket/src/main/java/com/proyectointegrador/msticket/service/implement/TicketServiceImpl.java
@@ -1,6 +1,10 @@
 package com.proyectointegrador.msticket.service.implement;
 
+import com.proyectointegrador.msticket.domain.PaymentMethod;
 import com.proyectointegrador.msticket.domain.Ticket;
+import com.proyectointegrador.msticket.exception.PaymentMethodNotFoundException;
+import com.proyectointegrador.msticket.exception.TicketNotFoundException;
+import com.proyectointegrador.msticket.repository.IPaymentMethodRepository;
 import com.proyectointegrador.msticket.repository.ITicketRepository;
 import com.proyectointegrador.msticket.service.interfaces.ITicketService;
 import lombok.RequiredArgsConstructor;
@@ -11,9 +15,10 @@ import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
-public class TicketService implements ITicketService {
+public class TicketServiceImpl implements ITicketService {
 
     private final ITicketRepository ticketRepository;
+    private final IPaymentMethodRepository iPaymentMethodRepository;
 
     @Override
     public Optional<Ticket> getTicketById(Long id) {
@@ -29,11 +34,18 @@ public class TicketService implements ITicketService {
 
     @Override
     public Ticket createTicket(Ticket ticket) {
+        PaymentMethod paymentMethod =
+                iPaymentMethodRepository.findById(ticket.getPaymentMethod().getId())
+                .orElseThrow(() -> new PaymentMethodNotFoundException("Payment method not found"));
+        ticket.setPaymentMethod(paymentMethod);
         return ticketRepository.save(ticket);
     }
 
     @Override
     public void deleteTicket(Long id) {
+        if(!ticketRepository.existsById(id)) {
+            throw new TicketNotFoundException("Ticket not found");
+        }
         ticketRepository.deleteById(id);
     }
 }
